// src/api/Learnup/models/HTMLTag.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export enum HTMLTag {
    PARAGRAPH = 0,
    HEADING1 = 1,
    HEADING2 = 2,
    HEADING3 = 3,
    HEADING4 = 4,
    HEADING5 = 5,
    UNORDERED_LIST = 6,
    ORDERED_LIST = 7,
    LIST_ITEM = 8,
    BLOCKQUOTE = 9,
    PREFORMATTED = 10,
    CODE = 11,
    HORIZONTAL_RULE = 12,
    STRONG = 13,
    EMPHASIS = 14,
    FIGURE = 15,
    FIGCAPTION = 16,
    TABLE = 17,
}